src/test/java/DSVToJsonLConverterTest.java
import static org.junit.Assert.assertEquals;

import org.example.DSVToJsonLConverter;
import org.junit.Test;

import java.io.IOException;
import java.util.List;

public class DSVToJsonLConverterTest {

    @Test
    public void testConversionWithFile1() throws IOException {
        // Provide the path to the first test input file
        String inputFilePath = "input/DSVinput.txt";

        // Expected JSON-L output for the first file
        String expectedOutput = "expected JSON-L output for file 1";

        // Perform conversion
        DSVToJsonLConverter converter = new DSVToJsonLConverter();
        String actualOutput = converter.convertDSVToJsonL(inputFilePath);

        // Assert the output matches the expected JSON-L output
        assertEquals(expectedOutput, actualOutput);
    }

    @Test
    public void testConversionWithFile2() throws IOException {
        // Provide the path to the second test input file
        String inputFilePath = "path/to/second/input/file.txt";

        // Expected JSON-L output for the second file
        String expectedOutput = "expected JSON-L output for file 2";

        // Perform conversion
        DSVToJsonLConverter converter = new DSVToJsonLConverter();
        String actualOutput = converter.convertDSVToJsonL(inputFilePath);

        // Assert the output matches the expected JSON-L output
        assertEquals(expectedOutput, actualOutput);
    }

    @Test
    public void testConversionWithMockData() throws IOException {
        // Define mock data for testing
        String mockData = "mock DSV data here";

        // Expected JSON-L output for the mock data
        String expectedOutput = "expected JSON-L output for mock data";

        // Perform conversion
        DSVToJsonLConverter converter = new DSVToJsonLConverter();
        String actualOutput = converter.convertDSVToJsonL(mockData);

        // Assert the output matches the expected JSON-L output
        assertEquals(expectedOutput, actualOutput);
    }

    // Add more test cases as needed to cover different scenarios

}
